requirements.txt
asgiref==3.8.1
Django==4.0.10
djangorestframework==3.13.1
legacy-cgi==2.6.1
pytz==2024.2
sqlparse==0.5.3
